🏗️ Architect: Modularized apts/equipment/plotting.py
Decomposed the EquipmentPlottingMixIn "God Object" into a specialized package structure:
- apts/equipment/plotting/base.py: Shared internal utilities and core bar chart plotting logic.
- apts/equipment/plotting/simple.py: Specific plots for magnification (zoom) and field-of-view (FOV).
- apts/equipment/plotting/graph.py: Complex connection graph visualization logic using networkx.
- apts/equipment/plotting/__init__.py: Recomposed the public EquipmentPlottingMixIn and maintained backward compatibility for test mocks.

This improves maintainability by following the Single Responsibility Principle and separating concerns between simple chart generation and complex graph layout logic.
